+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @opengeni/core
 
+## 0.11.4
+
+### Patch Changes
+
+- Updated dependencies [7c6aa7c]
+  - @opengeni/config@0.7.2
+  - @opengeni/db@0.12.2
+  - @opengeni/documents@0.2.32
+  - @opengeni/runtime@0.13.6
+  - @opengeni/storage@0.2.26
+  - @opengeni/events@0.3.23
+
+## 0.11.3
+
+### Patch Changes
+
+- Updated dependencies [d03ee4b]
+  - @opengeni/runtime@0.13.5
+
 ## 0.11.2
 
 ### Patch Changes